improve region forward
